Rewrite shortcut code to handle accelerators for Gio::Actions.
Changes include:
* Converting to C++ class.
* Removing duplicate code.
* Removing dead code.
* Fixing memory leak.
* Eliminating improper use of GDK_MODIFIER_RESERVED_XXX values.

To do:
* Handle window and document level Gio::Actions
* Integrate verbs and actions in preferences.

+++ src/ui/dialog/inkscape-preferences.h
@@ -514,7 +514,7 @@ protected:
         Gtk::TreeModelColumn<Glib::ustring> id;
         Gtk::TreeModelColumn<Glib::ustring> shortcut;
         Gtk::TreeModelColumn<Glib::ustring> description;
-        Gtk::TreeModelColumn<unsigned int> shortcutid;
+        Gtk::TreeModelColumn<unsigned long long int> shortcutid;
         Gtk::TreeModelColumn<unsigned int> user_set;
     };
     ModelColumns _kb_columns;
